+++ CHANGELOG.md
@@ -29,7 +29,8 @@
 ### Added
 - Histogram with fill distribution on Score popover
 - Browse scores in Score popover
-- Shell and Node scripts for updating scorelist
+- Development tool for updating scorelist
+- Popover can be opened from diagram
 
 ### Changed
 - Boardsize parameter on configuration screen